server/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
flask_sqlalchemy
psycopg2-binary
python-dotenv
flask_cors
requests
requests-oauthlib
gunicorn
psycopg2-binary
flask_session